simplify UnmarshalSSZ and move BeaconBlockIsNil
